refactor: change styled components to ts files and update imports
