+++ src/components/fixtures/Header.tsx
@@ -1,7 +1,7 @@
 import { MobileNavigation } from "./MobileNav";
-import { HeaderWrapper } from '../styled/LayoutWrappers';
-import { DesktopNav, DesktopNavLink, DesktopNavList, DesktopNavItem } from '../styled/Header/DesktopNavigation';
-import { HeaderBrandTitle, HeaderContent } from '../styled/Header/HeaderBaseStyle';
+import { HeaderWrapper } from '../styled/global/LayoutWrappers';
+import { DesktopNav, DesktopNavLink, DesktopNavList, DesktopNavItem } from '../styled/fixtures/DesktopNavigation';
+import { HeaderBrandTitle, HeaderContent } from '../styled/fixtures/HeaderBaseStyle';
 import { HeaderLogo } from './PageLogo';
 
 /* ---- Header layout ---- */
