+++ src/components/styled/fixtures/PageLogo.ts
