+++ src/components/styled/fixtures/footerStyle.ts
