+++ src/components/styled/global/Pages.ts
