+++ src/components/styled/home/TopOccupationsStyle.ts
