+++ src/components/styled/jobdetails/JobDetailsPageStyle.ts
