+++ src/components/styled/jobs/SearchBarComponents.ts
