+++ src/components/styled/jobs/ShowJobsComponents.ts
