+++ src/pages/HomePage.tsx
@@ -1,15 +1,15 @@
-import { PageSection, PageTitle } from '../components/styled/Globals/Pages'
+import { HiddenPageTitle, PageSection } from '../components/styled/global/Pages'
 import Hero from "../components/hero/Hero";
-import heroImg from "../components/styled/hero-job.jpg";
+import heroImg from "../assets/hero-job.jpg";
 import TopOccupations from '../components/TopOccupationsBar'
 
 export const HomePage = () => {
 
   return (
     <PageSection>
-      <PageTitle> Home Page</PageTitle>
+      <HiddenPageTitle> Jobcentralen</HiddenPageTitle>
       <Hero
-        imageSrc={heroImg} 
+        imageSrc={heroImg}
         imageAlt="Laptop och kaffekopp"
         title="JobbCentralen"
         quote="Hitta jobbet du trivs med — låt jobbet hitta dig."
