+++ src/pages/JobsPage.tsx
@@ -3,8 +3,8 @@ import { SearchBar } from "../components/SearchBar";
 import { ShowJobs } from "../components/showJobs";
 
 import { Paginator } from "../components/Paginator";
-import { HiddenPageTitle, PageSection } from '../components/styled/Globals/Pages';
-import { BaseBlockWrap } from '../components/styled/Globals/Wrappers';
+import { HiddenPageTitle, PageSection } from '../components/styled/global/Pages';
+import { BaseBlockWrap } from '../components/styled/global/Wrappers';
 
 
 export const JobsPage = () => {
