+++ src/pages/PageLayout.tsx
@@ -4,7 +4,7 @@ import { useReducer } from "react";
 import { JobContext } from "../contexts/JobContext";
 import { jobReducer, initialJobState } from "../reducers/jobReducer";
 import type { JobsLoaderData } from "../loaders/jobsLoader";
-import { FooterWrap, MainWrap, PageWrap, Typography } from "../components/styled/LayoutWrappers";
+import { FooterWrap, MainWrap, PageWrap, Typography } from "../components/styled/global/LayoutWrappers";
 import { Header } from "../components/fixtures/Header";
 import { Footer } from "../components/fixtures/Footer";
